Changed neuro and GCS valid values.
The valid values specified in the selection attributes of the fields were integers that were being used directly to calculate the score. This would make any change to the valid values or scores difficult in future, possibly needing database migration scripts to change all '1's to '3s' for example. Decoupling the value from the score it represents should minimise any refcatoring required if the values change in the future.

2 letter codes were used as the values for the fields to align with the 'NT' and 'NO' values that are specified in the story, and also to save on space in the database. Using separate labels for each field value gives us the luxury to make use of small values like this.

Also added tests and other bits.