Labels changes.
Ensured that ' Observation' is appended to the observation names in the tasks view.
Ensured that 'Daily ' is prepended to the food and fluid name in the data visualisation and take observation lists.
Ensured that 'bare' observation names are shown for all other observations in the data visualisation and take observation lists.