Re-added deepcopy of form description.
At some point the `min` key of the weight definition is being changed to a `str`, which causes validation to fail later. Doing a deepcopy means that whatever bit of logic converts to a `str` is doing so on a copy, and so doesn't affect future uses of the form description.