Rapid tranq (#69)
* Working tests and implementation, other tests now broken.

* More fixing of tests.

* Fixed more tests.

* Fixed tests.

* Fixed flake8 error.

* Fixed flake8 error properly.

* EOBS-1141 fix.

* Fixed flake8 error and Codacy issue.

* EOBS-1146 fix. No need to override just to get new PMEs on the report anymore.

In order to fix the original bug, it was most sensible to have all the PMEs retrieved in one go, rather than in bits, with each module overriding and adding it's own new PMEs. The PME events are now created dynamically by looking for any models with the pattern `nh.clinical.pme.*` and creating events from them. There is no need to override in individual modules just to get new PMEs on the report anymore.

Fixing the bug revealed another one: the events are not shown in chronological order. The events are created by processing a list of activities, each of which can potentially produce 2 events (a start and a stop event) which aren't necessarily one after the other chronologically. This means the list of events produced is fundamentally out of order, so a sort has been added to the list of events.

* EOBS-1153 and 1154 fix.

- Changed 'rapid tranquilisation' label in mobile UI to be upper case and therefore consistent with the desktop label.
- Was incorrectly assigning cancel reason to the non-existent `cancel_reason` field, instead of `cancel_reason_id` field. Therefore the cancel reason was not set and was not picked up by the report data logic.
- Added an `__init__.py` to the nhclinical repo as it must be a Python package in order to run flake8 / Pylint to run against it, allowing you to check the entire repo at once.
- Added some test coverage.