EOBS-1259 fix. (#76)
Added a keyword argument to the `reformat_server_datetime_for_frontend` method to also convert the datetime to the client's timezone. Unfortunately the API is a little awkward in that it wants a context to be passed. The reason for this is that the `datetime_utils` model itself has no context of it's own and therefore no timezone, so the caller must pass their context which is then monkey patched onto the `datetime_utils` model.