Refactoring adt admit model.
We have a register model as well as a patient model which represent the registering of a patient in a hospital. This is for auditability and to mirror the clinical concepts you see in PAS systems that send ADT messages. We should therefore never create a patient record without creating a patient record either. This feature aims to address this by locking down creation of patients and making models (such as admit) depend on the adt register model instead of the patient model.

This particular commit attempts to swap out the `patient_id` field on the adt admit model for a `registration` field that references the adt register model. There is still more work required to test and build a migration script that will transfer interpret the old `patient_id` data into new `registration` ids.