Bug fix, still need to add regression test.
Even if you delete a value after adding it and submit, Odoo's UI still records that the field was changed in some way and submits the new field value to the server. In the case of this bug it is an empty list of ward IDs, but the logic on the server was a bit over zealous in that it would complain as soon as it saw any value at all being passed for the ward IDs without checking whether there were actually any IDs in that value (its just an empty list in this case).