Added usage of registration in more places.
Replaced the usage of `patient_id` in the `nh.clinical.adt.patient.admit` model's submit and complete methods, now they use `registration` field instead. Amended tests as necessary.

In the `complete` method another admission record of a different model, `nh.clinical.patient.admission`, is created. This record also depends on `patient_id` rather than a register record but we are not changing this, so in the `submit` method we look up the `patient_id` from the `registration` field and pass it to the creation of the other admission record. Confusingly there is not a second 'register' model type like there is for admission so even if we wanted to we could not change the second admission model to also depends on a register model rather than the patient model.