Updating API method implementation.
Now that the `nh.clinical.adt.patient.admit` field has changed other areas of the code must be updated. One of the main places is the `nh.eobs.api` which is called by the ADT connector. The ADT connector should be unaffected as the method signature has not changed, only its implementation.