Fixed migration applying same ID to every row.
Before removing the patient ID column its values are used to populate the newly created registration column. The problem was that all the rows were being updated with a single value rather than dynamically. Unfortunately the fix for this was not elegant as an UPDATE query is now run for each individual row. It should be possible to do this in a single query but the same problem kept recurring and it was no longer worth the time investment just to have a more elegant way of achieving the same result.