Fixes for EOBS-1735, 1739, and 1740.
Searching was not showing correctly. Had to override `name_search` on the register model.

Name field was being populated on the register records due to the `default_name` key being added to the context. Added a condition to check for and delete that key in the `create` method override in the register model.

Added unique constraints on hospital number (other_identifier) and nhs number (patient_identifer) to match the constraints on the patient model.