Fixed tag used for our Odoo fork.
The `VERSION` environment variable is used for cloning our fork of the Odoo repo during the tests which is then used to test against.
There are tags in the repository and some of them have different `requirements.txt` entries. One of these entries is a URL that is no longer accessible.
Changing to the newer tag means there is `Python-Chart==1.39` in the requirements rather than a stale URL.