Fixed and uncommented some test modules.
The `test_doctor` module was already working. According to an old commit this was commented out because it was logging warnings and it was suspected that was contributing to a Travis failure, the test was never uncommented after that.
The `test_devices` and `test_spell` modules just need a minor fix to their patient creation to get them working again. They were originally commented out for the same reason as `test_doctor`.