Fixed some fields not being validated.
Some fields weren't being validated against their configured minimums and maximums because they were not all being passed explicitly to the validation method. This commit makes the validation of fields more managed with less burden on the developer. Now the only place developers have to explicitly list field names is in the constrains decorator arguments. Will look into a way to manage this also but no solution was obvious due to the fact that the class reference does not appear to be accessible in the decorator (PyCharm was saying the reference could not be resolved).