Fixed reallocation changes not being persisted.
The bug was caused because we migrated to a later version of Odoo V8 which moved the creation of the transient record that backs a wizard to submission rather than on opening of the wizard. Before this change the created record was passed back to the server with the changes despite the allocating field being read-only. Now read-only fields are omitted from the call to creation so the server does not get them. The fields should never have been read-only if they are intended to be updated so removing read-only attribute was part of the fix that ensured the server got the correct data.

An unintended side-effect of this was that a delete icon now showed on all the 'allocating' rows due to a different bug in Odoo's web addon. Had to add a dirty override of one of the widgets in order to make it honour the `delete` attribute in the views which could then be used to switch that functionality off (as can already be done with other widgets).