Added extra unit tests to cover missing cases.
The method expects there to be either an allocation or reallocation transient record. This happens a lot in test cases because shift changes are often created but in reality most of the time there will not be any transient records because they will have been vacuumed. This means that logging in, opening the reallocation wizard and trying to do an allocation fails.