Removed warning message.
Because it was only created with the assumption that it should never in theory be triggered. It was intended to be a 'just-in-case' message if there is a bug. However Nayira pointed out that it can still be triggered normally during a shift change. That means that it is likely to occur and is therefore damaging to the user experience so it is being removed.